add namespace cv::sphereview3d
